sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fairops-sdk"
version = "0.1.0"
description = "FairOps SDK — 3-line ML bias monitoring integration"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Toro Bees", email = "team@torobees.dev"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic==2.6.0",
    "google-cloud-pubsub==2.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.0",
    "pytest-cov==4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
